+++ src/py/mat3ra/notebooks_utils/pyodide/ui.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Optional
+from typing import Any, List, Optional
 
 from ..primitive.prompt import create_prompt_text, get_integer_from_input
 
@@ -27,32 +27,3 @@ async def ui_prompt_select_array_element_by_index_pyodide(
     result = array[index]
     print(f"Selected {element_name}: ", array[index])
     return result
-
-
-async def select_coordination_threshold_emscripten(distribution: Dict[int, int], default_threshold: int) -> int:
-    """
-    Select the coordination threshold from the given distribution. Works in Pyodide environment.
-
-    Args:
-        distribution: The distribution of coordination numbers.
-        default_threshold: The default threshold value.
-
-    Returns:
-        int: The selected coordination threshold.
-    """
-    coordination_threshold = default_threshold
-    coordination_numbers = list(distribution.keys())
-    prompt_text = f"\nCoordination numbers distribution: {distribution}" f"\nEnter coordination threshold value: "
-    while True:
-        try:
-            value_str = await input(prompt_text)  # type: ignore
-            value = int(value_str)
-            if value in coordination_numbers:
-                coordination_threshold = value
-                break
-            else:
-                print(f"Invalid value. Please enter one of these coordination numbers: {coordination_numbers}")
-                break
-        except ValueError:
-            print(f"Please enter a valid integer value from: {coordination_numbers}")
-    return coordination_threshold
